requirements.txt
aiogram~=2.14
environs~=8.0.0
pillow~=10.2.0
APScheduler~=3.10.4
schedule~=1.2.1
python-dateutil~=2.9.0.post0
requests~=2.31.0